Use TOML based configuration
This commit abstracts the configuration for catch and capture into a
TOML file on romfs. I had some trouble with virtual functions not being
called and the conversion to c strings causing pointer errors. The code
is therefore a bit messy. I'll clean this up once I transition to using
multiple implementation and header files (yes, I'm going back on going
back on that decision).

To parse the TOML I added an additional dependency on cpptoml. I chose
TOML because it's easy to read and familiar to most (since it's like
.ini). I avoided json because jansson seems to be finnicky and
nlohmann/json also caused issues for others. XML is bad and YAML has
unneeded complexities.

The new approach isn't yet implemented for interactive mode because that
requires a proper menu system. The whole thing doesn't work on 3dsx with
CMake due to a bug in 3ds-cmake where the romfs doesn't get included for
3dsx targets.

Closes #3; Closes #4; Closes #9 (not atomic, sorry).